Modulo9_TestesUnitarios/requirements.txt
langchain
langchain-openai
python-dotenv
pytest